+++ src/vs/workbench/contrib/issue/browser/issue.contribution.ts
@@ -10,11 +10,13 @@ import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle
 import { IProductService } from 'vs/platform/product/common/productService';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { Extensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
-import { WebIssueService } from 'vs/workbench/services/issue/browser/issueService';
-import { IWorkbenchIssueService } from 'vs/workbench/services/issue/common/issue';
+import { WebIssueService } from 'vs/workbench/contrib/issue/browser/issueService';
+import { IWorkbenchIssueService } from 'vs/workbench/contrib/issue/common/issue';
 import { BaseIssueContribution } from 'vs/workbench/contrib/issue/common/issue.contribution';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from 'vs/platform/configuration/common/configurationRegistry';
+import { IIssueMainService } from 'vs/platform/issue/common/issue';
+import { IssueFormService } from 'vs/workbench/contrib/issue/browser/issueFormService';
 
 
 class WebIssueContribution extends BaseIssueContribution {
@@ -26,6 +28,7 @@ class WebIssueContribution extends BaseIssueContribution {
 Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench).registerWorkbenchContribution(WebIssueContribution, LifecyclePhase.Restored);
 
 registerSingleton(IWorkbenchIssueService, WebIssueService, InstantiationType.Delayed);
+registerSingleton(IIssueMainService, IssueFormService, InstantiationType.Delayed);
 
 CommandsRegistry.registerCommand('_issues.getSystemStatus', (accessor) => {
 	return nls.localize('statusUnsupported', "The --status argument is not yet supported in browsers.");
