+++ src/vs/workbench/contrib/issue/electron-sandbox/issueMainService.ts
